Added saccade peak velocity metric to output file 2 created in ED.py
The second output file (Output file 2) now reports the peak velocity for a given saccade. This is done by searching for the max velocity values in the orignal list that this created to calculate saccade amplitude for each saccade. This measure will be helpful when trying to investigate metrics for workload transitions